model Texture and Animations was Changed
Added geckolib dependencies to the build.gradle file to work animations on the model.
Also added new files -- movement animation model, new chester model and texture.
These files was charged into ChesterModel class.
In ChesterEntity class was implemented IAnimatable to add the animations with the respective file.
Finally the Eye_Bone item was configured to spawn ChesterEntity on the world.